// lib/tours/rendering/avatars/tour-avatar.types.ts
export type HeyGenAvatarAlphaAnalysis = {
  sourceWidth: number;
  sourceHeight: number;
  sampledFrameCount: number;
  alphaThreshold: number;
  medianBox: VisiblePixelBox;
  maxBox: VisiblePixelBox;
  transparentPadding: AvatarTransparentPadding;
  edgeTouchRate: AvatarEdgeTouchRate;
  cropRisk: AvatarCropRisk;
};

export type VisiblePixelBox = {
  x: number;
  y: number;
  width: number;
  height: number;
  right: number;
  bottom: number;
};

export type AvatarTransparentPadding = {
  left: number;
  right: number;
  top: number;
  bottom: number;
};

export type AvatarEdgeTouchRate = {
  left: number;
  right: number;
  top: number;
  bottom: number;
};

export type AvatarCropRisk = {
  level: "none" | "low" | "medium" | "high";
  reasons: string[];
};
// ...
export type HeyGenAvatarFrameCheck = {
  timestampSeconds: number;
  path: string;
};

export type HeyGenAvatarWorkflowWarning = {
  code: "avatar-edge-touch" | "avatar-too-tight" | "frame-check-failed";
  message: string;
  severity: "info" | "warning";
};
// ...
export class TourAvatarError extends Error {
  constructor(
    message: string,
    readonly code:
      | "MISSING_HEYGEN_API_KEY"
      | "MISSING_HEYGEN_AVATAR_ID"
      | "MISSING_AVATAR_SOURCE"
      | "HEYGEN_CREATE_FAILED"
      | "HEYGEN_POLL_FAILED"
      | "HEYGEN_DOWNLOAD_FAILED"
      | "AVATAR_ALPHA_ANALYSIS_FAILED"
      | "AVATAR_LAYOUT_FAILED"
      | "AVATAR_VIDEO_UPLOAD_FAILED"
      | "AVATAR_METADATA_UPLOAD_FAILED"
      | "AVATAR_VIDEO_ASSET_CREATE_FAILED"
      | "AVATAR_METADATA_ASSET_CREATE_FAILED"
      | "FRAME_CHECK_FAILED"
  ) {
    super(message);
    this.name = "TourAvatarError";
  }
}